fix(singleuser): should not print pwd if it is specified
